refactor: move http handle constructor to its own module
